Add files via upload
Task 1
The "Bank Marketing Data Set" from the UCI Machine Learning Repository is related with direct marketing campaigns (phone calls) of a Portuguese banking institution.
The classification goal is to predict if the client will subscribe a term deposit (variable y). You can find a description of the attributes at the original UCI URL, https://archive.ics.uci.edu/ml/datasets/Bank+Marketing.

Task 2
The following is a database of 1698 Hindi Movies from 2005-2017: https://www.kaggle.com/datasets/rishidamarla/bollywood-movies-dataset
A movie is a hit if revenue > budget, and it is a flop otherwise. The goal is to predict whether a movie will be a hit or flop, given all the other attributes.